rust fix: defer initialization of statics
The `ver` and `messages` static variables are initialized with strings
in the C version, which become `(*b"...").as_ptr()` in the Rust version.
But `as_ptr` is not a `const fn`, so this is not a valid constant
expression.  The fix is to initialize the statics with `0 as *mut u8`,
and call a function before `_c_main` to overwrite them with correct
values.