Added options/colours to test script
There are now command line options to choose which tests to run,
what log level to set (by default, we log nothing), and the output
is helpfully colourized.

Including the word "fail" in the first line of a test case marks it
as being expected to fail (which means it doesn't show up as RED in
the terminal output.