Add back Paren type, add extra checks in conversion
In 814baddb83671f22356d461173342923baf298ac I had removed the paren
type from the C AST and I was merging the paren node with its
contents. This didn't work because we had already visited the type
inside the paren, so it was too late to try to merge.

The moral of the story is that you can _only_ merge nodes when you
are certain you haven't already visited it.

  * added comments about merging nodes
  * added some assertions in 'conversion' so that this sort of thing
    doesn't go unnoticed again
  * added back the Paren type