Actually get the right initializer for VarDecl's
Global declarations like the following now have an initializer:

    struct { struct { int l; } j; } k;

The trickiness is around tentative definitions. I _think_ we can
safely assume a VarDecl is a defintion only if it isn't a declaration.
We already only visit canonical top-level decls so if the canonical
top-level decl is a tentative definition, I think that means it is
a definition.