Move into the loop body small blocks
Any blocks which are reached by exactly one block in the loop body and
which have no successors should be moved into the block.

One common case of this is `return` blocks. These are always leafs in
the CFG, often with only one predecessor. We don't want these to end up
in a goto_table after the loop - we want to return from inside the loop.