Move badly-placed declarations to top scope
The idea is simple: when building up structures, whenever you place a
basic block that claims a certain variable is in scope, and that
variable is not in scope, mark the declaration associated with that
variable.

Then, all marked declarations get moved to the top of the function in
a second pass.

Most of the challenge is making sure the right information gets where
it needs to be:

  * we build `CFG`s and `Structure`s of `StmtOrDecl`, not just `Stmt`
  * `Structure`, `StructureLabel` are now also parametrized over stmt
  * we store information about how to translate decls when building the CFG
  * we store information about which decls to lift when building structures
  * finally, we flatten out this indirection right before simplifying structures