Mutability cast in volatile/reads writes
One common optimizable edge case for volatile reads/writes is
reading/writing from/to a pointer you just dereferenced. I already
optimized that to generate the `read_volatile`/`write_volatile` call
with the initial pointer.

However, we still need to add a cast if the pointer doesn't have the
right mutability. This is what I added in this commit.

Also added support for `--translate-entry` for main functions which
return void.