Disable any sort of name-shadowing in Relooper
This is a bummer, but I can't think of any reliable way to shadow names properly.
The issue is that we choose names when building the CFG but we only really find
out the scopes (when we could determine if we need to rename) when we reloop the
CFG. Instead, we'll disallow any sort of shadowing.

Fixes #102.