Nicer codegen + semantic fix of stmt expr
Now that we check that statement expression CFGs don't have return stmts,
we know that the only return must be the last one. That means that when
processing the output it is safe to match on the last statement and, if it
is a return, we can make a simple expression block instead of a closure
which gets called immediately.

Also, closures definitely have to be by reference, not `move`ed. Derp.
Test updated.