rust-refactor: wrap_api: fix up function pointers after changing ABI
`wrap_api` changes the ABI of each function it wraps from `extern "C"`
to the standard Rust ABI.  This doesn't cause any problems for direct
calls to the function, but it changes the type of any pointers to the
function.  `wrap_api` now updates function pointers to refer to the
generated wrapper (which keeps the old `extern "C"` ABI) instead of the
wrapped function.