ast-importer: fix some cases of overly aggressive decl pruning
`c_ast::iterators::DFNodes` previously failed to traverse into the
`CQualTypeId` children of certain `CExprKind`s.  Specifically, it failed
on some `CExprKind`s where the initial `CQualTypeId` field represents a
type explicitly written in the C program (which `DFNodes` should
traverse) and not just the inferred result type of the expression (which
it should skip).

Fixes #160