rust-refactor: I made a ton of "improvements" to
the code which were both for readability and optimization.
Thus meaning that there was a lot of stuff that I found to be a bit
convoluted, that needed to be removed/changed. Paths are changed
higher than before, and there are a lot less AST traversals.
Still a lot to be done.