The `reorganize_modules` transform moves nodes
around, but there may be a time when there is a node within a module
with an Ident `unnamed`, and an incoming node with the same Ident but
is a completely different type.
```
mod foo {
    mod foo_h {
        // going to be moved into `foo`
        struct unnamed {
            a: i32
        }
    }
    struct unnamed {
        b: u32
    }
}
```
As seen, the naming should be a bit more verbose. Which is the reason
for this commit.
```
mod foo {
    mod foo_h {
        // going to be moved into `foo`
        struct unnamed_foo_h {
            a: i32
        }
    }
    struct unnamed_foo {
        b: u32
    }
}
```