rust-refactor: This is a fix for nested use
statements that have `super` as the prefix. The issue was that they
were not getting incorporated at all, what this fix does is separate
each segment into it's own simple use statement. It is inserted into
the module only if there is no other occurrence of the use statement
within the module already.