literate: add new options for building tutorials
This adds options that are useful for tutorials and similar
documentation, where code to refactor is included inline in the
Markdown, not in a separate .rs file.

The new options are used by applying attributes to fenced code blocks.
For example:

    ```rust refactor-target hidden
    fn main() { /* ... */ }
    ```

This is a block of Rust code (`rust`), which will be used as the crate
being refactored for future `refactor` blocks (`refactor-target`), and
will not be displayed in the output (`hidden`).

This also adds some global refactoring options, which can be set using a
`refactor-options` code block:

    ```refactor-options hidden
    auto-revert
    diff-style = full
    ```

This enables the `auto-revert` option (`no-auto-revert` would disable
it), and sets the `diff-style` option to `full`.

Since the new options can now require multiple `c2rust refactor`
invocations for a single input file, this commit also includes some
refactoring of the code for invoking `c2rust refactor` and processing
the outputs.