refactor: always run rustc in non-incremental mode
In incremental mode, the compiler panics when we try to re-analyze the
same crate.  Since cargo turns on incremental compilation by default, it
was previously necessary to set `incremental = false` in the
`Cargo.toml` when running multi-step refactorings with `--cargo`.  Now
we override `session.opts.incremental` during refactorer startup, so the
`Cargo.toml` changes are no longer needed.