Update the Travis script to run on Linux and OS X
In this commit, the `.travis.yml` (Travis build script) has been
updated to build and test on both Linux and OS X. To achieve this
the `script` portion of the script, checks for the os type and runs
the correct provisioning script accordingly. Another change is, `matrix`
was added, and this sets the correct environment variables depending on
the os.