Update the Travis script to run on Linux and OS X (#36)
* Update the Travis script to run on Linux and OS X

In this commit, the `.travis.yml` (Travis build script) has been
updated to build and test on both Linux and OS X. To achieve this
the `script` portion of the script, checks for the os type and runs
the correct provisioning script accordingly. Another change is, `matrix`
was added, and this sets the correct environment variables depending on
the os.

* Update the linux provisioning script

* Add a missing dependency to the provision script

Added `python3-setuptools` to the linux provisioning script,
because the travis xenial box does not seem to install recommended
packages. Also moved `os: linux` into `matrix:` within the travis
script.

* Test `cargo build` version of the project

* Try to build both ways concurrently

* Forgot `then` after the if statement

* Try switch statement for `CACHE_NAME` check

* Specify brew installed `llvm_config` path for OSX build

* Move environment variable to matrix

* Rename `CACHE_NAME` environment variables

* Remove variables from global scope

* Condense switch statement to reduce redundancy

* Move some comments to a proper position

* Fix a small nitpick in a comment

* Extremely minor comment change