Do not run the refactor tool automatically
On most occasions the refactor tool should not be ran after translation,
because it could fail. Some of the reasons are: the `rename_unnamed`
command may need to be ran before, some unnecessary files can get
translated and fail to compile (files containing vararg functions), or
the `reorganize_definitions` command does not work for very large
projects yet.