Add Lua refactoring script interface.
Lua scripts can now be executed by the refactoring tool to control refactorings. We plan to build out the API as needed.

This commit also moves functionality in api.rs to other modules and removes the glob use

Glob imports of the api module cause problems for code analysis tools like
racer, and transforms were pulling in a lot of functionality not exposed in the
api module anyway. Removing the api module and moving functionality to other
relevant modules paves the way for a new clean, external API in a new api module
that we can expose over RPC for external refactoring scripts.