Remove shallow from .gitmodules
Git submodules are locked to a specific commit. If that commit is not HEAD and
shallow is true, the checkout can't find the commit it's looking for since only
the most recent commit was fetched. We can't do shallow checkouts here because
we need to pin to commits that may not be in the shallow fetch.