Always invoke `rustc --print sysroot` the same way
Probing for the existence of rustup's home directory is unreliable,
because for example someone might have used it in the past but not have
it installed now. Also, it's complicated to do.

Instead, arrange that whether rustup is in use or not, the same `rustc`
command will always produce the answer we wanted. This works by passing
the desired toolchain in an environment variable, which rustup respects
but rustc ignores, instead of as a command-line argument which rustc
doesn't understand.

Thanks @thedataking for the initial fix! It worked for me but I hope
this alternative version will be more reliable and easier to maintain in
the future.