Ensure executability for all Python with a __main__
Neither of the scripts I needed to run to reproduce the robotfindskitten
results was executable, which strikes me as an easily-eliminated point
of friction, so here's a patch to do so.

The way I did it had the side effect of making the interpreter lines
identical across all of the scripts. (Some of the existing scripts
included an extra, but harmless, space.)

For each Python script with a __main__ block, I ran each of these three
steps:

- delete existing interpreter if any: `sed -i '2,$b; /^#!/d'`
- insert interpreter: `sed -i '1i#!/usr/bin/env python3'`
- set executable permission bit: `chmod a+x`

To apply those steps to the right files, I prefixed each command with:

git grep -lFw __main__ '*.py' ':^:**/__main__.py' | xargs