Don't strip comments during relooping
The relooper was stripping comments off of statements by overwriting an already
set span. I don't think we need to combine new comments into an existing comment
span after the first pass of relooping because all comments should already be
attached to a statement. If this invariant is false, we will need to extend the
CommentStore to merge comments into an existing comment span.