Track source ranges for each C AST element for comment handling
Comments which fall into a source range, e.g. of a function, but are not
attached before any other AST elements should be attached to the end of the Rust
AST element they are contained in. This change makes that happen, instead of
attaching trailing comments to the next Item after it.