Fix rewriter issue with multiple inserts at the same point
Inserting multiple items at the same point in a module would try to replace the
same whitespace chunk multiple times and fail. Replacing whitespace would be
nice here, but it we can't reliably do this. At some point I'd like to
reintroduce this but only for the first new item insert at the rewrite location.