Add --preserve-unused-functions (#291)
* Add --preserve-unused-functions

to keep otherwise pruned static / inline functions.

Closes: https://github.com/immunant/c2rust/issues/287

* transpile: Reduce code duplication

* prune_unused_decls: Rename to unwanted

The criterion for whether it is desired for an item to be transpiled
have been broadened to possibly preserve unused functions as well. Thus,
the "unused" name part has been replaced with "unwanted", which
traditionally means unused but, in other configurations, can mean
"unused but not a function" now as well.