Add entry and exit point hooks
We need to hook the entry and exit points of the program to ensure that
event logging is set up correctly and that all events have been flushed,
respectively. This instrumentation assumes that the program being
instrumented exits either normally or by unwinding. Calling
proccess::exit() will result in events being lost.