[WIP] Support va_lists in structs
This patch is limited to handle a (subset of) a pattern that can
be found in the apache httpd server to parse scripts. More general
support for va_lists in structs is difficult to implement since
it forces us to propagate lifetimes to non-local uses of such
structs.

This also adds support to the Builder interface for inserting
lifetime generic parameters and for building lifetimes, which was
missing before.