test: support architecture-specific transpiler tests
architecture-specific tests are denoted by ending their name with .$arch
(such as .x86_64) and including a "target-tuple" file containing a target
tuple on which the test is known to work.

running tests on a host of the same architecture works as before; if the
architectures differ, the target tuple is used to correctly transpile the
test and (if an appropriate toolchain is present) compile the transpiled
sources as well. lacking this toolchain is not considered a test failure.