Add instrumentation cargo support and c2rust command
Switches instrumentation to building with cargo rather than requiring a
complete rustc invocation. The analysis runtime will be built at the
same time to ensure it is built with the same rustc version as we are
basing the instrumentation on.

Adds a `c2rust instrument` command to drive this instrumentation build.