Fixed the logic in `c2rust-ast-exporter` for asserting only one source path is being evaluated at a time.
Previously, it assumed sources would accumulate in
`OptionsParser.getSourcePathList()`
and not be cleared out between successive calls,
so it maintained a static counter to check against.
But in clang 13 or 14, this changed,
and now `OptionsParser.getSourcePathList().size() == 1`,
so the assert would fail whenever `process` was called multiple times,
or when `c2rust_transpile::transpile` was called multiple times.

This changes the code to handle both cases.
That is, either `OptionsParser.getSourcePathList().size() == 1`,
or `OptionsParser.getSourcePathList().size() == source_path_count`.

I tested this on clang/llvm versions 9 through 14.