Added cross compilation capabilities to `test_translator.py` using `qemu` for emulation and `cargo-zigbuild` and `zig cc` for cross-compilation.
Specifically, ...

`cargo-zigbuild` is used in place of `cargo`
for building and running the test crates.
It uses `zig cc` as the linker, which has seamless cross compilation.

`qemu` is used to emulate other targets and architectures.
These are small test programs, so this should be fine,
and doing it this way allows all tests to be run locally,
which is very convenient.

I wasn't sure how dependencies should be managed,
but for this, I just had them installed in `test_translator.py` itself,
assuming it's running Ubuntu Linux.
If there's a better way to do this, let me know,
but I just did it inline first to make things quick.

Specifically,
* `cargo-zigbuild` is installed using `cargo install`
* `zig cc` is installed through `pip install ziglang` (added it to `requirements.txt`)
* `rustc` targets are installed through `rustup target add`
* `qemu` is installed through `apt` (`qemu-user`, `qemu-user-static`)
* `qemu` dynamic libraries are installed through `apt` (`gcc-${triple}`)
* `update-binfmts` (for enabling `qemu` `binfmt_misc` support) is installed through `apt` (`binfmt-support`)