Fixed compile errors in `tests/longdouble`, which isn't normally tested.
I'm still unsure how to transpile this test,
since `c2rust transpile` panics on `"1.3" is not a valid Ident`,
and I think I need special things set up to test this.

But the rest of the test crate also had issues.
The pinned git version of the `f128` crate didn't compile anymore,
so I switched to the latest version on `crates.io`, which seems to work.