Switched `tinycbor` to use `GIT_REPOSITORY` and `GIT_TAG` instead of `URL` and `URL_HASH` in `ExternalProject_Add`.
`tinycbor` is already downloaded from GitHub, so we can just use the built-in `git` support.
This makes upgrading `tinycbor` versions easier.

Also, we can use either git tags/branches or commit hashes.
I used a commit hash here so we keep the integrity checks (which are SHA-1 now vs. MD5),
but we could use git tags for more flexibility.

This also has the advantage over raw tarball hashes in being more stable,
as GitHub can (and has) changed the compression slightly, resulting in a different hash.
Using the commit hash is from the source and thus should be completely stable.