translator: remove `std_or_core()` function
This is a redundant complication of translator source code: the `libcore` is always available,
whether the crate is `#![no_std]` or not. Using the imports from `core` or `std` is purely a
matter of code style, which should not be the responsibility of the transpiler.

If the transpiled crate should use `libstd` instead of `libcore` in imports, this can be fixed
by some refactoring tools, or possibly even `rustfmt`.