Added a `$BINARY` option to `pdg.sh` that if absent is calculated by `get-binary-names-from-cargo-metadata.mjs`.
This allows us to optionally avoid needing `node` (like in CI),
while still allowing it to be used locally for convenience for now.

I could also replace the scripts with a Rust one using https://docs.rs/cargo_metadata
(which does the heavy lifting, and precisely, so it'd be a simple program),
though that adds yet another Rust crate that will be temporary,
as the need for these scripts will soon be obviated once https://github.com/immunant/c2rust/issues/448 is fixed.

Thus, I'm just adding a workaround for CI for now.