When pretty-printing a `Graph`, instead of the `GraphId` index, print the `Graph`'s hash.
The hash (a `u64`) is turned into a `u32` and then printed as base64 to be succinct.
The reason we print the hash instead of the index is so that `Graph`s remain stable when their position changes,
so when changes cause a new `Graph` to be emitted,
it doesn't cause a snapshot diff on all subsequent `Graph`s, only the new one.