Support mutable references in most cases.
The Rust borrow checker uses two-phase borrows
in cases where mutable references are used as
arguments to a method. This poses a challenge
for the dynamic instrumentation in cases where
inserted statements activate the first phase of
a two-phase borrow. So, addressing this special
case will be deferred to a later time.