Added a new implementation of `c2rust-instrument` that invokes `cargo` as a subcommand, overriding `RUSTC_WRAPPER`.
More specifically, instead of using `cargo` as a dependency,
which is enormous and causes some compilation issues (https://github.com/immunant/c2rust/issues/488),
we invoke `cargo` as a subcommand.
This also has the advantage of using the correct `rustup` `cargo` wrapper so that it resolves correctly.

Then we set `RUSTC_WRAPPER` to our own binary.
We use `RUSTC_WRAPPER` to detect if we're being invoked as a `cargo` or `rustc` wrapper.

If we're the `cargo` wrapper, we parse args using `clap` derive,
determine the crate target info (that we want to instrument)
using `cargo metadata` instead of the massive `cargo` dependency,
and then run `cargo` as a subcommand,
passing the `cargo` args as is (so all normal `cargo` invocations all work),
our own executable as `$RUSTC_WRAPPER`,
and the crate target info and instrumentation args serialized through `$C2RUST_INSTRUMENT_INFO`.

If we're the `rustc` wrapper, we parse the instrument info,
determine if we should instrument the current `rustc` command
if the crate being compiled matches the crate target info passed,
and based on that, either invoke `rustc`
or run `RunCompiler::new(...).run()` with our `MirTransformCallbacks`,
and then save the metadata to the metadata file.

Note that we have to pass the `--sysroot` to `RunCompiler`
as normally `rustc` looks this up relative to its executable's location,
which works when the `rustup` `rustc` wrapper resolves to the toolchain-specific `rustc`,
but since we're using `RunCompiler` directly and being `rustc`,
we need to explicitly set the sysroot.

Instead of passing `c2rust_analysis_rt` as an `--extern` and running `cargo` separately to compile it,
we just add it as a normal dependency to the crate we're instrumenting.
In this way it's similar to `libc`, which can be `extern`ed as it's in the sysroot,
but the preferred way is to specify it as a normal dependency.