Now that we use `cargo` normally and include `c2rust_analysis_rt` as a normal dependency, we can delete only the main binary, forcing it to be rebuilt and re-instrumented.
Most of the time building used to be in building `c2rust_analysis_rt`,
and on `donna`, was ~15 seconds for `c2rust instrument`.
Now it takes only ~0.5 seconds, 30x faster.

Thus, we now also unconditionally instrument.
Before we only did when the `c2rust-instrument` binary changed,
but this missed changes to the instrumented crate.
Now that it's fast enough, we just always instrument.