Don't instrument build scripts.
`lighttpd` was failing to be instrumented because we tried to instrument its build script, since `$CARGO_PRIMARY_PACKAGE` is set for it.
We can't instrument build scripts because `c2rust-analysis-rt` is not yet built for them,
and running the build script would then fail because environment variables needed by the runtime may not be set.
And we don't want to instrument the build scripts because they're at compile time.

To detect this, we check for when `--crate-type bin` is specified (using `rustc_driver`'s arg-parsing code)
and yet there's also no `$CARGO_BIN_NAME`, which is set for normal binary crates.