Remove the `cargo clean --package {root_package}` (not needed since incremental is disabled).
Thus, we have to handle truncating and appending to the metadata file differently,
because we have to handle two cases:
1. No inputs change and `cargo` doesn't run `rustc`.  In this case we should keep the existing metadata file.
2. `cargo` runs `rustc` with incremental disabled and we append to the metadata file.

Because of 1., we can't blindly truncate the metadata file.
Instead, we leave it as is, and if we find we've appended to an existing metadata file
(because the length increased), then we truncate the old part of it.