Support struct fields in static analysis.
This adds support for generating concrete Origins
for each struct lifetime parameter and resolving
lifetime parameters to those origins when visiting
the MIR body. It also adds support for ADT field
pointer permissions.

There are three stages to making this possible:
1. gather hypothetical (i.e. pointer-based) and actual (reference-based) lifetimes
2. generating concrete origins
3. resolving lifetime parameters when visiting the MIR body

Stage 1 runs in a loop and gathers ADT metadata. The loop terminates
when a fixed point is reached, as determined by changes in the ADT
metadata. This approach is to handle ADTs that are recursive and/or
mutually recursive.

Stage 2 generates one concrete `Origin` per ADT lifetime parameter,
per MIR local. When grabbing the `Origin` of a given struct field,
even nested ones, it is one of these concrete origins that will be
returned.

Stage 3 traces the concrete Origins generated for each ADT lifetime
parameter all the way to the last projection / field access. It does
this by updating a mapping for each projection element visited.