(`c2rust-analyze`) Rename `Callee::{MiscBuiltin => Trivial}` and expand it to any trivial library function.
The meaning of `MiscBuiltin`/`Trivial` is in that it requires no special handling, i.e. it is trivial.
Thus, `Trivial` is a much more appropriate name than `Misc`.

This also expands the definition of `Trivial` to non-builtins,
including any "library"/external function that takes or returns no pointers,
as they are all trivial.