(`c2rust-analyze/tests`) Save `c2rust-analyze` output during tests and print them if it crashes.
Previously, for `filecheck.rs`, we piped directly to `FileCheck` and asserted on `status.success()`.
And for `lighttpd.rs`, we just asserted on `status.success()`.
But when `c2rust-analyze` crashed, which happens often, we just got the exit code, but no error messages.

So now we always save the `c2rust-analyze` output to `{input}.analysis.txt`
and print it in case it crashes,
as well as the status and the command just ran (so it can easily be run independently).

By saving it to a file, this also has the benefit of avoiding
the `unsafe` used to duplicate the pipe to `FileCheck`.

This also moves most of the code in `mod common`, which is shared among integration tests,
as otherwise there'd be a lot of useless duplication.