(`c2rust-analyze`) Support `*c_void` casts from alloc functions and to free functions.
This is a rework of #788 that properly separates the two `*c_void` cast directions.
In #788, these are conflated.  I'm not exactly sure if this can lead to a bug
(the `Place`s in the map are not necessary unique),
but the point is, the conflation makes it harder to understand and reason about,
so it is harder to tell if there is one or not.
It may be correct due to details of our particular analysis and how MIR works,
but then you have to understand all the details to know if it's correct or not,
and it's a lot more brittle to future changes.

By separating the `From` and `To` directions here, it IMO makes it much easier to understand.
It allows the logic to be more easily separated, such as in
* `CVoidCastDirection::from_callee`
* `CVoidPtr::checked`
* `CVoidPtr::get_cast_from_stmt`
* `CVoidCasts::should_skip_stmt`

Separating the logic also allows each part to be more documented,
which IMO makes it a lot easier to readon about and understand.

It also separates the `*c_void` casts code into its own module.