(`c2rust-analyze/tests`) Changed `filecheck/trivial.rs` to never use currently unsupported features.
This mainly includes aggregate support, e.g. `()`, `None`, etc.
Generics are also not supported yet, so those can't be used.

Thus, `f!` now defines a function with the requested type already existing as a parameter,
and then calls itself to create a call with that type.

`!` checking is also now skipped as that's unsupported.