(`c2rust-analyze/tests`) Add `Trivial` tests for recursive/self-referential types through pointers.
The test for the same but through references is also added but commented out,
as it's not yet supported (#834) and will lead to infinite recursion and a stack overflow.
With recursion through pointers, this exits early with `false`, so it's not an issue,
and since transpiled types shouldn't have references, only pointers,
this shouldn't be an issue for a while until we get to partially-refactored code.