(`c2rust-analyze`) Split `Callee::Other` into `Callee::{Normal, UnknownDef}` and include `fn` ptrs in `Callee::UnknownDef`.
Currently, `Callee::Other` is as it sounds, just any other function
that is not first recognized to be `Trivial` or a builtin.
The docs for `Callee::Other` say it is statically known,
but `extern` functions with statically unknown definitions
are still included, and these currently panic
when we look up `LFnSig` in `visit_call_other`.

Also, calls to `fn` ptrs are completely unsupported,
but `lighttpd` does have a lot of them.

This tightens `Callee::Other` into what it says it is,
`Callee::Normal`, which is a normal (not any of the other special cases)
Rust function that has a statically known definition.

On the other hand, `Callee::UnknownDef` is added
for cases where the function's definition is not known,
and is meant to be treated as a potential fallback for any function call.
This includes `extern` functions, whose definition
may be completely unknown and dynamically linked.

And it also includes `fn` ptrs who may have multiple definitions.
It may be possible to statically determine
the set of function definitions it could be, but not always,
and doing so would be an optimization over this fallback.

Also, since we can now handle `FnPtr`s,
we also consider them for triviality checking.