(`c2rust-analyze`) Used `canonical_path` to massively simplify `builtin_callee` (now inlined into `ty_callee`).
This can also be done in other places, such as in:
* #788: for checking `*c_void` is the correct type.
* #852: for `core::{const,mut}_ptr::offset_from`

so it should be a good investment simplifying many things going forward as well.

This also skips the previously considered `[T; N]::as_(mut_)ptr` functions,
as they are not a thing.
`tcx.types.u8` is also used instead of creating a new `Ty` for `u8`.