(`c2rust-analyze`) Reformulate `Trivial`ity into the more correct and comprehensive `PtrEffect`s.
Instead of categorizing all function calls into trivial (`Callee::Trivial`) or non-trivial,
we now categorize only `Callee::UnknownDef`s into a scope of `PtrEffects`.

These new `PtrEffects` are also determined more correctly than the old `Trivial`ity,
which generally means they're more conservative,
though there are options for making assumptions that the code is not
overly adversarial and behaves somewhat nicely,
in which case the `PtrEffects` are greatly reduced in most cases,
which is important in keeping static analysis useful.

For example, we now categorize the `libc` function `close` (`unsafe extern "C" fn close(fd: i32)`)
as `PtrEffects::Deep` rather than `PtrEffects::None` because it is `unsafe` and `extern`.
For this reason, it's re-commented out in `lighttpd-minimal`.
When it comes to handling the `Callee::UnknownDef`, though,
there won't be a difference, as `i32` has no pointers contained deeply in it.