Only generate polonius facts for reborrows for Use(Copy)
Previously when generating reborrows for Use(Move), the moved
place would be borrowed where, in the original code, that wouldn't
be necessary, i.e. it's lifetime may have been shorter, leading to
polonius errors. For example, if the original MIR had a local moved
follow by a StorageDead for that local, the lifetime would have been
extended for that local and the StorageDead would no longer be
helpful if the LHS of the move was for e.g. written to at a later
time.