Returned labeled type U in `V = T as U`
As per the discussion [here](https://github.com/immunant/c2rust/pull/839#discussion_r1115160479),
type_of_rvalue was combining portions of types T and U and returning labeled type X. This
forced some similarity checking in the types that ideally should be placed in a
yet-to-exist is-castable-to check. This commit labels the type U and makes that the LabeledTy
for the RValue of each cast.