(`c2rust-analyze`) Add all `libc` `fn`s called by `lighttpd_rust_alalgamated` to `all_known_fns`, un-annotated.
There are a lot of `fn`s, so this just adds them unannotated so far.
Since `known_fns!` checks the number of perms matches the number of pointers,
new `fn`s with pointers are commented out for now.

A few `libc` `fn` names are included as comments
as they are for some reason not in Rust's `libc` crate, but are a part of `libc`.

The other foreign `fn`s in `lighttpd_rust_amalgamated` are `pcre2` `fn`s, not `libc` ones.