unlower: look for last terminator statement across multiple blocks.
Sometimes we want to `get_last_call` where the input `locs` span multiple
blocks. This means that sometimes the `last()` statement is not a call as it
would normally be in a list of locations constrained to just one block, so instead
we traverse the statements in reverse and return the first found terminator.