`c2rust-analyze`: Add a `cargo` wrapper so it can be invoked like `cargo` on a whole crate.
This copies the approach from `c2rust-instrument`.

The main difference are:
* We don't care about the instrumentation and metadata `c2rust-instrument` need.
* We still allow `c2rust-analyze` to be called as a `rustc_wrapper` directly.
  More specifically, the `cargo` wrapper is supposed to set `RUST_SYSROOT`.
  `c2rust-instrument` requires this, while `c2rust-analyze` will re-calculate it if it wasn't.

This allows us to keep using the `rustc` wrapper in tests,
as the tests are all set up as single files meant to be compiled with `rustc` directly.
We can change this, but that'll come later.
For full crates we test like `lighttpd-minimal`, we'll definitely switch to the `cargo` wrapper,
at the very least for a proof of concept.