Keep track of unique projection combinations in Project
Add an index value to Project events to differentiate between
different projections with the same pointer offset, e.g.,
(*p).x and (*p).x.a if a is the first field of the structure.

This is implemented as an IndexSet<Vec<usize>> where each element
is a unique combination of field projections. The Project index
points to an element in this set.