transpile: fix `curl` transpile breakage while maintaining a total `SrcLoc` order
#1128 fixed the non-total (non-transitive) `SrcLoc` ordering,
but accidentally broke the `curl` transpilation test in `c2rust-testsuites`,
which is tested in CI, but was broken for external contributors PRs (fixed now in #1159).

In the `curl` transpilation, a couple of `use` imports (`__sigset_t` and `C2RustUnnamed_4`)
were missing, cause the build to fail afterward.

I'm still not exactly sure why this fixes the issue while maintaining a transitive, total order,
but it passes the total order test and transpiles `curl` correctly now.
Hopefully this is a complete fix, and I didn't just fix a one-off error in `curl`.