Fix references to `rust-toolchain` without `.toml` suffix
Two .gitignore files and a GitHub workflow referenced `rust-toolchain`
without a `.toml` extension, but all the files in the repo and the
generated files are `rust-toolchain.toml`.

In addition, have the GitHub workflow's hashing take
`*rust-toolchain.toml` into account, to allow for a future split between
the toolchain for compiling c2rust and the toolchain for compiling
c2rust-generated code that needs nightly.