Move Linux CI from Azure Pipelines to GitHub Actions
This simplifies the CI setup substantially. In particular, this
installs packages as part of the CI setup, rather than in a separate
manual container build step where modifications by contributors are
ineffective until the container images are regenerated.

Now, changes to provisioning will take immediate effect, allowing
upgrades to new versions to be contributed and tested easily.

This does not move macOS over, because I don't have sufficient
experience with macOS to easily debug the issues that occur when
attempting to do so.

Fixes: https://github.com/immunant/c2rust/issues/1226