Allow running multiple transpiles in parallel
If c2rust-transpile needs to generate a temporary compile_commands.json,
previously it used a fixed path for it across all invocations. Instead,
create a uniquely-named temporary directory for it, so that parallel
transpiles don't clobber each other.

I've used a two-year-old release of tempfile to avoid adding a bunch of
duplicate versions of other crates to Cargo.lock, but presumably some
upgrades ought to happen sooner or later.