transpile: translate `__builtin_ia32_pause` as `core::hint::spin_loop` instead
`spin_loop()` is implemented as `_mm_pause()` (the `pause` instruction) on `x86`/`x86_64`,
but it's the safe and cross-platform version of it, so we prefer it.

Furthermore, while the x86/x86_64 `_mm_pause` is stable now,
the arm/aarch64 equivalent `__yield` is not and requires
`#![feature(stdarch_arm_hints)]` (`aarch64`) and `#![feature(stdarch_arm_neon_intrinsics)]` (`arm`).
`__builtin_ia32_pause` is obviously x86/x86_64 only, but I think it'd be better
to support the more cross-platform version that makes
implementing other architectures' equivalents a lot simpler.